# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Hugs as a Service API
  version: 1.0.0
  description: Platform for hugging other users with rate limits, balance, and ranks
servers:
  - url: /api/v1
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Gender:
      type: string
      enum:
        - male
        - female
    ErrorCode:
      type: string
      enum:
        - WEAK_PASSWORD
        - USER_ALREADY_EXISTS
        - INVALID_CREDENTIALS
        - USER_NOT_FOUND
        - CANNOT_HUG_SELF
        - COOLDOWN_ACTIVE
        - INSUFFICIENT_BALANCE
        - WRONG_PASSWORD
        - USER_BANNED
        - CANNOT_BAN_ADMIN
        - ALREADY_HAS_PENDING_HUG
        - HUG_NOT_FOUND
        - HUG_NOT_PENDING
        - HUG_EXPIRED
        - DECLINE_COOLDOWN_ACTIVE
        - PENDING_HUG_EXISTS
        - MAX_SLOTS_REACHED
        - USER_BLOCKED
        - INVALID_TELEGRAM_ID
        - TELEGRAM_ID_TAKEN
        - TELEGRAM_LOGIN_FAILED
        - HUG_TYPE_LOCKED
        - CANNOT_DELETE_ADMIN
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code: { $ref: '#/components/schemas/ErrorCode' }
        message: { type: string }
    User:
      type: object
      required:
        - id
        - username
        - role
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        role: { type: string, enum: [user, admin] }
        gender:
          $ref: '#/components/schemas/Gender'
        display_name:
          type: string
          nullable: true
          maxLength: 32
        telegram_id:
          type: integer
          format: int64
          nullable: true
        tag:
          type: string
          nullable: true
          maxLength: 20
    AuthResponse:
      type: object
      required:
        - token
        - user
      properties:
        token: { type: string }
        user : { $ref: '#/components/schemas/User' }
...
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
paths:
...
  /auth/check-username:
    get:
      summary: Check if a username is available
      operationId: checkUsername
      parameters:
        - name: username
          in: query
          required: true
          schema:
            type: string
            minLength: 3
            maxLength: 32
      responses:
        '200':
          description: Username availability result
          content:
            application/json:
              schema:
                type: object
                required:
                  - available
                properties:
                  available: { type: boolean }
        '400':
          $ref: '#/components/responses/BadRequest'
  /auth/register:
    post:
      summary: Register a new user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                  username:
                    type: string
                    minLength: 3
                    maxLength: 32
                    pattern: '^[a-zA-Z][a-zA-Z0-9_]*$'
                    description: 'Username must be 3-32 characters, start with a letter, and contain only letters, numbers, and underscores'
                  password:
                    type: string
                    minLength: 8
                    maxLength: 128
                    description: 'Password must be 8-128 characters, contain at least one letter, one digit, and one special character'
                  gender:
                    $ref: '#/components/schemas/Gender'

              additionalProperties: false
      responses:
        '201':
          description: User registered successfully
          headers:
            Set-Cookie:
              description: Refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          $ref: '#/components/responses/Conflict'
...
  /auth/telegram/init:
    post:
      summary: Start a Telegram login session
      operationId: initTelegramLogin
      description: >
        Creates a new Telegram login session. Returns a bot deep-link URL and a
        poll token. The client should open the bot URL and poll /auth/telegram/poll.
      responses:
        '200':
          description: Login session created
          content:
            application/json:
              schema:
                type: object
                required:
                  - bot_url
                  - poll_token
                properties:
                  bot_url:
                    type: string
                    description: Telegram bot deep-link URL to open
                  poll_token:
                    type: string
                    description: Token for polling the login result
        '503':
          description: Telegram bot not configured
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/telegram/poll:
    post:
      summary: Poll for Telegram login result
      operationId: pollTelegramLogin
      description: >
        Polls the status of a Telegram login session. Returns 202 while pending,
        200 with auth response on success, 403 if banned, 404 if expired/not found.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - poll_token
              properties:
                poll_token:
                  type: string
              additionalProperties: false
      responses:
        '200':
          description: Login successful
          headers:
            Set-Cookie:
              description: Refresh token cookie
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '202':
          description: Login session still pending
          content:
            application/json:
              schema:
                type: object
                required:
                  - status
                properties:
                  status:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...